+++ src/stories/DatePicker.stories.tsx
@@ -1,10 +1,11 @@
 import type { Meta, StoryObj } from "@storybook/react";
+import { ReactNode, useState } from "react";
 import DatePickerProvider from "../DatePickerProvider";
 import DaySlots from "../components/daySlots/DaySlots";
 import Header from "../components/header/Header";
 import Title from "../components/title/Title";
 import WeekDays from "../components/weekDays/WeekDays";
-import { TDatePickerProps } from "../types";
+import { TDatePickerOnChange, TDatePickerProps } from "../types";
 import { argTypes } from "./constants";
 
 const meta = {
@@ -23,7 +24,7 @@ const meta = {
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-const RenderDatePicker = <T extends boolean>(props: TDatePickerProps<T>) => {
+const Card = ({ children }: { children: ReactNode }) => {
   return (
     <div
       style={{
@@ -35,13 +36,15 @@ const RenderDatePicker = <T extends boolean>(props: TDatePickerProps<T>) => {
         padding: "10px",
       }}
     >
-      <DatePickerProvider
-        {...props}
-        onChange={(e) => {
-          props?.onChange?.(e);
-          console.log("onChange: ", e);
-        }}
-      >
+      {children}
+    </div>
+  );
+};
+
+const RenderDatePicker = <T extends boolean>(props: TDatePickerProps<T>) => {
+  return (
+    <Card>
+      <DatePickerProvider {...props}>
         <Title />
         <Header
           monthSelectStyles={{
@@ -62,7 +65,70 @@ const RenderDatePicker = <T extends boolean>(props: TDatePickerProps<T>) => {
         <WeekDays />
         <DaySlots />
       </DatePickerProvider>
-    </div>
+    </Card>
+  );
+};
+
+const RenderControlledDatePicker = <T extends boolean>(
+  props: TDatePickerProps<T>
+) => {
+  const [value, setValue] = useState(props?.initialValue);
+
+  const onChange: TDatePickerOnChange<T> = (e) => {
+    props?.onChange?.(e);
+    console.log("onChange: ", e);
+    setValue(e);
+  };
+
+  return (
+    <>
+      <Card>
+        <DatePickerProvider
+          {...props}
+          initialValue={value}
+          value={value}
+          onChange={onChange}
+        >
+          <Title />
+          <Header
+            monthSelectStyles={{
+              backgroundColor: "#f0f0f0",
+              color: "#000",
+              padding: "5px",
+              borderRadius: "5px",
+              outline: "none",
+            }}
+            yearSelectStyles={{
+              backgroundColor: "#f0f0f0",
+              color: "#000",
+              padding: "5px",
+              borderRadius: "5px",
+              outline: "none",
+            }}
+          />
+          <WeekDays />
+          <DaySlots
+            onClickSlot={(v) => {
+              console.log("onClickSlot", v);
+            }}
+          />
+        </DatePickerProvider>
+      </Card>
+      <div style={{ width: 300, margin: "auto" }}>
+        <button
+          onClick={() =>
+            setValue(
+              (props.isRange
+                ? [new Date("2024-06-24")]
+                : new Date("2024-06-24")) as TDatePickerProps<T>["value"]
+            )
+          }
+          style={{ background: "#ddd", padding: 7, borderRadius: 7 }}
+        >
+          Set Custom Value to 2024-06-24
+        </button>
+      </div>
+    </>
   );
 };
 
@@ -190,6 +256,29 @@ export const PersianInEn_US: Story = {
       dayFormat: "numeric",
       weekends: ["thursday", "friday"],
 
+      // yearRangeFrom: 1330,
+      // yearRangeTo: 1400,
+      // minDate: new Date("2024-01-01T00:00:00.000Z"),
+      // maxDate: new Date(),
+    },
+  },
+};
+
+export const ControlledComponent: Story = {
+  render: RenderControlledDatePicker,
+  args: {
+    isRange: false,
+    initialValue: new Date(),
+    calendar: "gregory",
+    config: {
+      locale: "en-US",
+      weekStartsOn: "monday",
+      showOtherDays: false,
+      otherDaysSelectable: false,
+      weekdayFormat: "short",
+      dayFormat: "numeric",
+      weekends: ["saturday", "sunday"],
+      weekendSelectable: true,
       // yearRangeFrom: 1330,
       // yearRangeTo: 1400,
       // minDate: new Date("2024-01-01T00:00:00.000Z"),